requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair
asttokens
attrs
azure-core==1.31.0
azure-cosmos==4.3.1
azure-identity==1.19.0
beautifulsoup4
blinker
Bottleneck
Brotli
cachetools
certifi==2022.12.7
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama
comm
contourpy
cryptography==43.0.3
cycler
debugpy
decorator
distlib==0.3.6
entrypoints
exceptiongroup
executing
Faker
favicon
filelock==3.12.0
fonttools
frozenlist==1.5.0
gitdb
GitPython
htbuilder
idna==3.4
importlib_metadata
ipykernel
ipython
jedi
Jinja2
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
lxml==5.2.1
Markdown
markdown-it-py
markdownlit
MarkupSafe
matplotlib==3.9.2
matplotlib-inline
mdurl
mkl-service
mkl_fft
mkl_random
more-itertools
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
nest_asyncio
numexpr
numpy
packaging
pandas
parso
pickleshare
pillow
pip-upgrade==0.0.6
platformdirs==3.2.0
portalocker==2.10.1
prometheus-client
prompt_toolkit
propcache==0.2.0
protobuf==4.25.3
psutil
pur==7.1.0
pure_eval
pyarrow
pycparser==2.22
pydeck
Pygments
PyJWT==2.9.0
pymdown-extensions
pyparsing
python-dateutil
python-dotenv==1.0.1
pytz
PyYAML
pyzmq
referencing
requests==2.28.2
rich
rpds-py
setuptools==75.1.0
six
smmap
soupsieve
st-annotated-text
stack-data
streamlit
streamlit-camera-input-live
streamlit-card
streamlit-chat
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-toggle-switch
streamlit-vertical-slider
tenacity
toml
toolz
tornado
traitlets
typing_extensions==4.12.2
tzdata
unicodedata2
urllib3==1.26.15
validators
virtualenv==20.22.0
virtualenv-clone==0.5.7
watchdog
wcwidth
wheel==0.44.0
win-inet-pton
yarl==1.17.1
zipp
